"6:05pm. Yeah, no point in resuming now. Oh the PL thread is finally up.
///

While in the background [spiking NNs are closing the gap](https://www.reddit.com/r/MachineLearning/comments/iv29mh/r_eventprop_backpropagation_for_exact_gradients/) with float-using ones, I continue to work on Spiral v0.2, eager not to be left behind by the flow of circumstances. In my last review I said I was 95% done with the typechecking phase. During the first half of the month I spent the time filling in the gaps and testing. In the type inferencer all the features that I wanted have been added, and the feedback from the hover provider works great. The bugs that I could find have been eliminated and the whole system is smooth and snappy. The errors get displayed directly in VS Code and the hover provider shows the type in the tooltip when I hover over variables. It is simply a thing of beauty. I have to give credit to the [Hopac library](https://github.com/Hopac/Hopac/blob/master/Docs/Programming.md) where it is due. Both it and [NetMQ](https://zguide.zeromq.org/) made the concurrency challenges much easier, and I would have had a lot harder time making progress without them. I was completely right to take my sweet time during the first half of the year to educate and build up my skills in concurrency.

Very recently, I've finished the partial evaluation prepass, and now I am gearing up to finish the language implementation. Once I do the partial evaluation pass, the language will be pretty much done. There is also the codegen after it, but out of all the passes the code generation is the easiest one in my view.

The partial evaluation pass is something I am deeply familiar with at this point and I have a large cache of code from the previous version, so it won't give me much trouble. Right now I am struggling though. It happens every time I decide to make a big step, the prepass was not an exception either - there is a big sense of inertia that I need to push myself to overcome. I need to coax, cajole and brainstorm; I need to shut myself away from the world and look inwards. Once I do that for long enough the strength needed to overcome the psychological inertial builds up and I am able to make progress.

Working a language is tough. As great as the finished, polished Spiral v0.2 will be, a great asset to both myself and the world at large, the language as it is now in its unfinished state is just a huge burden to me. Thanks to my newfound concurrency skills I feel like my programming has made a qualitative change and could act in a professional capacity regardless of the domain, but until v0.2 is done I will be completely chained to work on the language.

Though getting to a polished level will take me till the end of 2021, by the end of October I will have something that can compile code from start to finish and will be able to breathe welcome sigh of relief. My willingness to work on a compiler has been hitting my limit for a while now, and I would much rather program in Spiral instead. Though I am deeply familiar with partial evaluation and code generation, it is still a large pile of work I have to essentially grind though.

It's torture.

---

...On the bright side, in the 4-5 days since I've written the above paragraphs I've made absolutely remarkable progress on the partial evaluator. I've actually gotten all the hardest cases out of the way and at this rate I'll finish it in two days. Compared to how much work I had to do in order to get to a similar point back in 2017 and 2018, this rate of progress is like a plane. Just two more days of this, and then the codegen and I'll reach the finish line.

Another mountain of work will no doubt be waiting for me after that, but I'll have the leeway of choosing which mountain that will be rather than being railroaded as I am now. I am bursting at the seams to go from working *on* Spiral to working *in* Spiral.

///

Let me cut it off from 'It's torture.' and I'll write something else.

6:50pm. Done with lunch, let me finish the review.

7:05pm. This is good. Let me go with this.

Posted it. And I am done for the day here.

7:10pm. Tomorrow, I will be crushing it. After I finish the partial evaluator I'll slow down a bit before resuming with the codegen. After I am done with the codegen, there will still be plenty of work to integrate it with the editor, but that won't require reams and reams of code."