"5:35pm. I've settled on the design. There is no way around it. By necessity, the code generator will need to have a dependency on the partial evaluator. The way I will get around the problem of having nominals is by taking `ty_to_data` and returning it as a function along with the join point dictionaries.
```fs
let data_free_vars call_data = // ...
let data_term_vars call_data = // ...
```

Thankfully these two are globals.

5:45pm. Let me pull apply outside of term.

```fs
let peval (env : TopEnv) s x =
```

Now I can get rid of this `s` argument on the outermost level.

5:50pm.

```fs
    let s : LangEnv = {
        trace = []
        seq = ResizeArray()
        cse = [Dictionary(HashIdentity.Structural)]
        i = ref 0
        env_global_type = [||]
        env_global_term = [||]
        env_stack_type = [||]
        env_stack_term = [||]
        }
    let ty_to_data x = ty_to_data {s with i = ref 0} x

    term_scope s x, {|method=join_point_method; closure=join_point_closure; ty_to_data=ty_to_data|}
```

Actually, nevermind. I decided not to pull the apply. Instead I will just assume that the expression passed into peval is `main ()`. That is `main` apply by `()`.

```fs
cse = []
```

Let me make the cse this. This is generally a bad idea, but in this specific situation it is a great idea.

```fs
seq = null
```

Also let me make the `seq` null.

```fs
    let s : LangEnv = {
        trace = []
        seq = null
        cse = []
        i = ref 0
        env_global_type = [||]
        env_global_term = [||]
        env_stack_type = [||]
        env_stack_term = [||]
        }
```

Now if this gets used in any unanticipated way, I'll get a null exception right away.

Yeah, this is what the starting environment should be. As a matter of fact, I can make the arrays null as well.

```fs
    let s : LangEnv = {
        trace = []
        seq = null
        cse = []
        i = ref 0
        env_global_type = null
        env_global_term = null
        env_stack_type = null
        env_stack_term = null
        }
```

Now it is really illustrative what I am expecting the expression passed into `peval` to require.

6pm. No way around it. I might have bragged what good grasp on the codegen I have, but v0.2 is different from the previous versions. Now that nominals exist, I need a way of figuring out what free variables lie inside them and I cannot do that without evaluating them.

I am going to have too loosed the requirements on what kind of information I am capable of transmiting between the passes. Alternatively, I supposed I could always story this information inside the join point keys. But then I will have a problem as to what to do with layout types.

Just being able to turn regular `Ty`pes into `Data` is the easiest solution to the broad spectrum of problems.

The disadvantage now is that there is a non trivial dependency between the partial evaluator and the code generator, but unless there is a programming error this won't be a problem. The function I am passing should act like `Ty -> Data` for all intents and purposes.

6:10pm. Mmmmhhhh...I am feeling it.

Drilling into a side problem is an excellent of building motivation to tackle the main one. I think I have everything I need to get started on this tomorrow.

I am going to get a good chunk of the code generator out of the way tomorrow. There is no reason why I should not finish it over the next few days. I need to maintain this work attitude.

After that comes tying up the phases by making the server, and then testing. I'll finally be able to produce F# code from Spiral's.

That will be the start of everything. That is when the next phase will come. By the end of the month I will be able to seriously start thinking about getting my hands on neurochips. I've been putting thought into how to sell out properly and I'll be able to see how that will work out.

Let me stop here from the day. It is time to unwind. I've done my brainstorming and now it is time to put that aside. If I have to do more thinking, I should leave it for during the work day tomorrow."