"3:35pm.
```fs
    let file_message uri f =
        let tokenizer = tokenizer uri
        let parser = parser uri
        let typechecker = typechecker uri
        let hover,_ = hover uri

        Hopac.start (
            let res = IVar()
            Ch.give tokenizer (f res) >>=.
            IVar.read res >>= fun (blocks,tokenizer_errors) ->
            queue.Enqueue(TokenizerErrors {|uri=uri; errors=tokenizer_errors|})

            let req_tc = IVar()
            Ch.give typechecker req_tc >>=.
            Ch.give hover req_tc >>=.

            let res = IVar()
            Ch.give parser {blocks=blocks; res=res} >>=.
            IVar.read res >>= fun parser ->
            queue.Enqueue(ParserErrors {|errors=parser.errors; uri=uri|})

            let res = IVar()
            IVar.fill req_tc {bundles=parser.bundles; res=res} >>=.
            IVar.read res >>= fun tc ->
            tc |> Seq.foldJob (fun errors tc ->
                IVar.read tc >>- fun tc ->
                let errors = List.append tc.errors errors
                queue.Enqueue(TypeErrors {|errors=errors; uri=uri|})
                errors
                ) [] >>- ignore
            )
```

I've forgotten about this beastie. I even forgot that the parser does it all at once.

3:50pm. I am thinking whether I could trim some of the other servers. Nah.

4:05pm. I am taking a break here.

I've gotten a small taste of it. From what I can see, from here on out, all the challenge is essentially dealing with these servers. The prepass, and the rest of the phases I can all do as a single server.

That is what I will aim for now. Later on though, it will be more challenging once multiple files enter the picture.

Well, let me take a break for a while and I will see whether I feel like doing anything else for the day.

6pm. Done with lunch. I am still thinking about it.

By 'it' I mean packages. Within a single project the compilation is linear so dealing with it is not too difficult. But packages can be arranged in a graph like fashion.

6:05pm. This makes things greatly challenging to deal with straightforwardly.

6:10pm. Change of plans. I said I am going to focus on testing, but forget that. I'll leave it for last. What I am going to do right now is deal with compilation thoroughly. I'll light up the project files, bring in packages and multi file compilation.

I am starting to realize that there are certain rules and design guidelines I have to follow when doing concurrent programming.

What I have in mind right now seems fancy, but it is the most realistic way of dealing with things. It is not the usual kind of programming pattern that I am used to, but I'll have to make do with it.

6:25pm. Now I am really hung up on this. Whatever thoughts of leisure I previously had are firmly replaced by obsession now.

6:40pm. Let me just stop here as I've already started browsing on the side.

Tomorrow I will dedicate myself to dealing with project management. After that is done, the only thing left will be testing.

I said that I wanted to deepen my relationship with concurrency. Well, this work will be my chance.

It is time for the next step. Tomorrow, the onslaught will resume."