"9:15am. Yesterday was another day I went to bed at 8pm because I was so tired.
I kept thinking about packages. Unlike regular files which have a linear compilation order in the package, a package A can have packages B and C as dependencies. And B and C should be build concurrently rather than linearly.

One thing in particular that gave me trouble is how to communicate information to A that B and C have changes.

I wanted that multiple other packages could read the output of one, so that rules out channels. A channel is only suitable for one to one communication, otherwise I would need multiples of them.

I kept iterating it in my head: channel, observable, mvar, ivar of ivars, ivar of channels...

Basically I would really have like is something like a MVar for which the others can react to when it changes. This was a huge problem for me that I could not figure out how. That I got to the root of it, and imagined, since I am already thinking of ivar of ivars, what if I added another field that signifies changes?

In the end I imagined something like this:

```fs
type 'a T = T of 'a IVar * 'a T IVar
```

Yeah, this was exactly what I needed. This would work for communicating information to multiple parties. What is this?

A fucking stream.

So in the end I literally thought of everything except the tool that I needed. When I made the connection, it came me the last of all.

Hopac has streams. It is just that I was too dumb to use them.

9:30am. Getting more intimate with them will be my job today. The reason this happened is because I've cemented the image of streams being infinite lists in my head. This is wrong; what I should be thinking instead is MVars that can be reacted to.

One thing I am going to get to the bottom of today is how to create promises. I keep passing those IVars everywhere, but that implies that other parties can fill them. Instead I should clearly indicate what is input and what is output.

```cs
  /// <summary>Represents a write once variable.</summary>
  public class IVar<T> : Promise<T> {
```

Ahhhhhhh...so IVar inherits from Promise. I thought of checking for this yesterday, and it turned out to be true!

9:35am. Well, that will make turning `IVar`s into promises straightforward. I just have to upcast them.

9:40am. Yeah, a plan is coming to me how to redesign all the servers so they use streams instead of channels.

```fs
        Hopac.start (
            let res = IVar()
            Ch.give tokenizer (f res) >>=.
            IVar.read res >>= fun (blocks,tokenizer_errors) ->
            queue.Enqueue(TokenizerErrors {|uri=uri; errors=tokenizer_errors|})

            let req_tc = IVar()
            Ch.give typechecker req_tc >>=.
            Ch.give hover req_tc >>=.

            let res = IVar()
            Ch.give parser {blocks=blocks; res=res} >>=.
            IVar.read res >>= fun parser ->
            queue.Enqueue(ParserErrors {|errors=parser.errors; uri=uri|})

            let res = IVar()
            IVar.fill req_tc {bundles=parser.bundles; res=res} >>=.
            IVar.read res >>= fun tc ->
            tc |> Seq.foldJob (fun errors tc ->
                IVar.read tc >>- fun tc ->
                let errors = List.append tc.errors errors
                queue.Enqueue(TypeErrors {|errors=errors; uri=uri|})
                errors
                ) [] >>- ignore
            )
```

I mean just look at this. I have these `res` variables here. I have to create them, fill them and then read them out.

I should just turn the errors into streams and put them on a conveytor belt straight to the editor. The same goes for the hovers. The way this works now is so complicated.

9:45am. I am going to redesign this all from scratch. But right now let me chill. Before I get to that thought, let me do something.

```fs
let tokenize text = (spaces >>. many_array token .>> (eol <|> tab) |>> Array.collect List.toArray) {from=0; text=text}
```

It has been bothering me how on an tokenizer error, the line will just get cut off.

```fs
let inline many_resize_array a d =
    let ar = ResizeArray()
    let rec loop () =
        let s = index d
        match a d with
        | Ok _ when s = index d -> failwith "The parser succeeded without changing the parser index in `many`. Had an exception not been raised the parser would have diverged."
        | Ok x -> ar.Add x; loop()
        | Error er -> if s = index d then Ok ar else Error er
    loop ()
```

I've been trying to refactor this last time, but failed. But I got the idea last night how to abstract this.

```fs
let inline many_iter f a d =
    let rec loop () =
        let s = index d
        match a d with
        | Ok _ when s = index d -> failwith "The parser succeeded without changing the parser index in `many`. Had an exception not been raised the parser would have diverged."
        | Ok x -> f x; loop()
        | Error er -> if s = index d then Ok() else Error er
    loop ()
```

This is what I should have done.

```fs
let inline many_resize_array a d =
    let ar = ResizeArray()
    match many_iter ar.Add a d with
    | Ok() -> Ok(ar)
    | Error er -> Error er
```

This is the way to do this.

```fs
let tokenize text =
    let ar = ResizeArray()
    let er = match (spaces >>. many_iter ar.Add token .>> (eol <|> tab)) {from=0; text=text} with Ok() -> [] | Error er -> er
    Array.collect List.toArray (ar.ToArray()), er
```

Yeah, this is the form I've been looking for.

```fs
let toks, ers = Array.map (tokenize >> function Ok a -> a, [] | Error er -> [||], er) edit.lines |> Array.unzip
```

This is how it was in tokenize previously. This will be the easiest refactor ever.

Let me give this a try.

...Yeah, it works just the way I'd want it to. Great.

I knew this pattern, and yet it completely slipped my mind. How embarassing.

10am. Things are looking up. My image of streams up to now was that they are immutable, but now that I know this is not the case - it should have been obvious if I thought about it, I have a lot more flexibility.

I can do this.

10:05am. I am going to have to decicate some time for this thought. Now way can I redesign all the servers that easily.

10:10am. One thing I still haven't figured out is how to deal with compilation triggers. Everything I can think of feels so complicated. Especially when I get to packages...

But I've made progress. Rather than start with packages right away, let me deal with multi file compilation first. That will be my first aim.

One way or another, I am going to get my wish of getting more familiar with Hopac. This is the time to really dive into it.

For now, let me watch Assault Lily and get breakfast. Then I'll do this properly."