"8:30am. I got up 5m ago and am already wasting my time reading Twitter posts.
In the background I am thinking, thinking, Thinking!

I am refining my design of the completed editor server. I think I have a bunch of very good ideas. It is all coming together for me.

9:25am. I'll take it easy today. I think the design is forming in my mind. Not just for multi file compilation, but the whole thing. Once this feeling subsides, all I need to do is commit to it and work to get it out.

What I am going to accomplish here is not going to be a small thing.

9:45am. Let me just slack for a while. I'll get breakfast right here.

I said I would take a break once I am done with the codegen, and I might as well execute that option. I want to dwell on it for a while longer. After that I'll start the run. Once I start, I won't stop until I finish the whole of the compilation harness.

That will be a major accomplishment.

Once I have that, Spiral will essentially have all it needs to function properly and all that will be left is to practice using it. I'd say testing, but testing never really ends. It is the steady accumulation of practice and fixing of bugs along the way that will bring it to completion.

I just have to do this one aspect and Spiral v0.2 will really be complete in the truest sense. Right now I have all the passes, and I just need to make the interface.

Things like GUIs and language servers are by no means a trivial thing.

A lot of concurrency experience is needed for both and I do not have much of that. But I'll get the practice as I go along."