"4:35pm. Let me move to the testing. I am really getting the feel for this now. Since none of the piecer are particularly troublesome I should be able to make steady progress from here on out.
Let me test this thing though.

```fs
| PatDefaultValue | PatFilledDefaultValue | PatValue | PatSymbol | PatB | PatE -> term
```

Ah, forgot about these.

For fuck's sake, F# is really making me work here.

Why did this pattern suddenly stop working? This should teach me a lesson about reporting bugs and upgrading the IDE.

4:50pm. Finally put wildcards everywhere.

Let me finally run this thing.

4:55pm. From what I can see, it seems I got everything right the first time around. Quite an accomplishment.

Project file parsing works as well.

```fs
let hover (req : (VSCPos * (string option -> unit)) Stream) (req_tc : TypecheckerRes Stream) =
    let req, req_tc = Stream.values req, Stream.values req_tc
    let rec waiting = req_tc ^=> processing
```

I'll make this an recursive object. I doubt F# is good at inlining this anyway.

5pm. I am doing some last minute checking up. Right now everything is perfect.

The program feels fine in my mind. Unlike before, the individual pieces are modular and my new principles are proving their worth.

With this I am back to par in terms of functionality.

5:10pm. I am thinking. At this point, it is only now that I can start seriously consider doing multi file complation. Because these functions are so modular courtesy of streams, I'll be able to reuse their functionality.

I can build on this. And that is a huge thing to me. This refactor was well worth it.

5:15pm. Obviously, I am not going to be starting anything new today. I'll leave this for tomorrow. But I won't call the day here. Let me think about the next step for a while longer.

I'll charge up my determination.

As soon as I thought this, the new goal came to me. That would be to finally take notice of the project file. The project file is the first step. I need the current project schema to be present.

Previously, for the past half week I mean, I've only been imagining the data flow at various levels. This was important, and it is what led to this radical simplification using streams to be a success, but up to now I've not really given project files more than a passing thought.

Obviously though, before I can do the multi file compilation I should get the file order. That is central.

Well, I could just assume it and do the project files last, but really, the right order to do things in is in what you expect to give you the most trouble.

It is best to make some distance from what I've been doing now.

I should do the project files.

An idea occurred to me.

```
dirOut: qwe
dirSource: c:\qwe
version: 1.2.3 (alpha)
name: Test
files:
    A
    Submodule/
        QQ
        QQ
        W
        E
        Main
    Main
```

For the stuff in files here, one thing that would be good is to add links to them directly in the editor. That way the user could have an easy time moving between them.

Yeah, this direction is actually a very good one to go in. Make the project files more interactive!

I am already typechecking them - I can go even further than this!

5:30pm. I am going to have to put in efforts to ensure various things. For example that only one package owns a file.

How will I find the project file once a `.spi` or `.spir` file is opened?

5:35pm. I know that other languages allow source directories, but to make things easier on myself I'll add a simple rule that ownership is based on envelopment.

If a project file if present in a directory then all the `.spi`/`.spir` files in that directory and subdirectories (not owned by others) belong to the project file.

That will be an easy way of doing things.

I want file ownership not to depend on the project file's contents, only its location.

I'll go with such a rule.

5:50pm. Ideas are really coming to me. I could use code actions to create missing files in project files. Renaming as well. That way a project file could be everything I need to manage Spiral projects. No need to use the file explorer to select files and going back and forth when fixing their order.

I can have it all in one place. That would be wonderful.

This is worth doing. I am going to do this.

6pm. It is time I put effort into this. I must do more than just get Spiral to work. I must go beyond what I did last time.

Doing Spiral project files in a polished manner goes hand in hand with doing multi package and multi file compilation. I have the skills and the knowledge that I need to do this.

The first half of 2020 was all about helping me overcome my shyness towards frontend development and concurrency. In the last few days I even took a step further by realizing the true value of Hopac's streams. For the next few weeks, I should go forward and get the full grade. I'll do this perfectly and move to testing the partial evaluator with a serene heart.

6:05pm. If I can do this work, I'll be able to move into seeing what the market demand is for Spiral as early as December. This could allow me to make income off it as early as January of next year.

I want to have a somewhat finished product published on the VS Code extension marketplace. When I get to that point, that is when I will know I am ready.

6:15pm. Let me stop here for the day. Tomorrow, I am going to start drilling into the project files. I'll deal with multi file compilation early on before I start putting in fancy features.

But the fancy features won't be hard at all though. I should be able to make good progress if I can stay focused. I've overcome my mental blocks regarding how to do concurrent compilation so making that a reality should not be far fetched."