"2pm. Let me do the chores here and I will resume.
2:05pm. Actually, let me chill a bit more.

I am thinking of what I have to do next.

2:10pm. Ownership validation and packages right? That won't be hard. And after I do that, the project files will have all of their functionality.

There is actually more to package management like this - I'd really want something like Nuget to store Spiral packages, but I've decided to leave that for later. I still do not know what the primary backend will be and so I do not want to sink time into this right now. Manual package management is fine for now.

Even the completed v0.2 will just be a demo ultimately. When the language has only a single user, I do not want to work on this. Later if I can garner some interest in this project, I'll place proper attention on this.

2:15pm. Let me do the chores. Let me just do this thing today, and I will finally be done with project files.

I'll be able to move to working on the compilation pipeline after that. And after that finally comes doing real programming in Spiral. I will go through the testing phase. I will put Spiral through its pace and after that, it will be a real tool worth using.

2:30pm. Done with chores.

Let me turn off the Internet. I do not want any distractions in case I get bored.

Now...

Ownership validation for files is what I need. But rather than files, it will be easier to do for folders. So that is what I will focus on.

2:45pm. This sort of thing will be a piece of cake.

3:20pm. I am thinking. Validating file ownership will in fact be easy, but checking for package circularity is a good bit harder.

If the user has many packages open, I do not want this to end up being a n^2 operation. So I am thinking how to do this efficiently.

3:55pm. I am completely sumped. Damn graph cycles. It is easy enough to figure out when a graph has a cycle. But what would be an efficient way of calculating when a node in a graph leads to itself. This is ridiculously tough. I've encountered this problem in the past under various guises.

Let me take a break here. After that I will Google this problem. I just can't figure out a O(log n) way of doing this. But there should be one somewhere.

4:15pm. This thing is going to ruin my whole day it seems.

https://en.wikipedia.org/wiki/Strongly_connected_component

4:25pm. Yeah, this stuff is what I need. Let me read up on this. Today at the very least I'll do the ownership validation. Let just get this out of the way first.

https://www.youtube.com/results?search_query=kosaraju%27s+algorithm&page&utm_source=opensearch

I do not get the Wiki article. Let me watch some of the vids here.

4:45pm. https://youtu.be/5wFyZJ8yH9Q
Kosaraju's Algorithm - Strongly Connected Components | GeeksforGeeks

I get this. This is pretty magical. This might be good to use in the prepass. Still, is there a way to do this incrementally?

```
There are also efficient algorithms to dynamically track the components of a graph as vertices and edges are added, as a straightforward application of disjoint-set data structures. These algorithms require amortized O(α(n)) time per operation, where adding vertices and edges and determining the component in which a vertex falls are both operations, and α(n) is a very slow-growing inverse of the very quickly growing Ackermann function.
```

https://cs.stackexchange.com/questions/96424/incremental-strongly-connected-components

This is going to be complicated. In order to give good error messages, I can't just say that the package has a cycle. I need to keep track of the strongly connected components.

https://en.wikipedia.org/wiki/Dynamic_connectivity

http://web.stanford.edu/class/archive/cs/cs166/cs166.1166/lectures/17/Small17.pdf

5:15pm. https://www.researchgate.net/publication/333644634_Connectivity_in_Directed_Graphs_Static_and_Dynamic
Connectivity in Directed Graphs - Static and Dynamic

Here is a 204 page thesis on the problem.

5:25pm. Let me read it for a while. I admit, I did not think that doing cycle detection along with good error messages would be this troublesome.

My progress has ground to a halt. I need to famliarize myself with the stuff in this thesis. I guess I'll focus on this problem for a while.

5:50pm.

```
We begin by presenting new algorithms for computing the maximal 2-edge- and 2-vertex-connected subgraphs in directed graphs in O(m3/2 ) time, where m and n are respectively the number of edges and vertices of the graph.
```

I need to keen in mind what `m` and `n` are.

...What is single source reachability?

6:25pm. Ok, I've pretty much wasted my entire day thinking about this problem. There are dynamic connectivity structures for undirected graphs, but I can't find anything on directed ones. In the thesis there is a structure for doing efficient edge removal, but I need a way to deal with the full problem.

I'll give up on doing it incrementally, and just use Kosaraju's algorithm.

6:30pm. Today I've thought about it for a few hours myself, realized the problem is really hard, then I looked around and in fact the problem really is hard.

I am going to curtail my ambitions for making a trully incremental package server. I'll have to recalc the entire tree on every update. Maybe if I packages from a central repo I could guarantee that they have no cycles and avoid expensive checks that way.

6:35pm. Forget that. Let me stop here for the day.

Today I finally made project files fully interactive which is a great accomplishment, but got stuck on packages. Tomorrow I will finish the validation, and go into the DFS algorithms more in depth. I want to study them before I commit."