"9am. So yesterday was as usual till 9pm, but then I started stuying stock charts for the next 2.5 hours again.
There are some things that will remain etched in my memory. Russia, emerging markets and China in the 00s. Short volatility after the 2008 crash. The IBB move in the 10s. US tech stocks after the crash in general. That last one is still ongoing.

I mean, just how stupid I am? Put up a monthly chart of the things I've talked about and you will see huge, persistent upward trends. Even child could have bought and held this. Yet, I could not. When I tried it with real money for two years between 2012-2014 I did jack despite studying it for 4 years before that. I learned jack.

The reason I could not win in the market was purely because of my lackluster character. My desire pushed me towards seeking an perfect system, so I ignored all the good ideas during that time. That was my undoing. I let the perfect become the enemy of the good.

9:10am. Take any of those charts, look at the dailies and you will see large moves against the trend. Because I sought perfection, and because I sought to protect my own time, I lived day to day instead of month to month. Because of that whenever there was danger I ran, and wasted my energy.

9:15am. Take a look at IBB during the period of 8-5 years ago. I was there, and I observed how resilient it acted during downturns past its breakout points. I conjectured it would be a good buy, but something like buying and holding an ETF position of 3 years - I never could have made such a move.

Imagine if I was present during the early 00s bear market. I could have observed the Russian market's incredible resilience during that period, but even if its behavior in September 2001 was a screaming buy, I could never have held on.

9:20am. Past 2014, I did not need any great insight on how to get better as a trader. I only had to reflect on my behavior for a while.

Right now, what are the huge points of interest? In one of the previous entries I mentioned long `QQQ`, but now that I've actually studied the market for a while, I'd definitely leave those aside and focus on more niche areas. `ARKK` is a definite interest of mine.

The power move though, would be to buy into `TAN` or `ICLN`.

Right now, the market has risen for a long while. For all I know it might top and start a new bear market. Buying right now would definitely be buying closer towards the end of the move than towards the beginning. It would be coming in late.

But even if `TAN` does not keep going up like crazy, it might flatline here and just whip around for the next 3-12 month, it won't just turn around and reverse its gains. That sort of thing - of powerful moves being reversed simply does not happen in the stock market. The same goes for `ARKK` here.

9:30am. My experience buying in situations like this is that I get a lot of volatility and not much gains, but if it were me right now, I'd buy `TAN` or `ICLN` - just to prove that I can hold for months now. That I can sit out of the volatility if there aren't obvious signs that I should get out of the market like in 2008 (finance stocks were getting crushed quarter to quarter in that period.)

9:35am. If the market really looks scary, I'd sell and wait a month or two. If things look fine, I'd buy back in, even if - in fact especially if, the vehicle I am interested has move higher than where I sold.

I just desire it. I just desire it so much. I have everything I need to make solid gains in the market. I have a system that finally makes sense to me. Long term, there are piles of money to be made from this.

I just do not have the money to participate in it right now.

But I do not regret that. Me not being able to trade are just the vagaries of life.

I do regret my weak character that lead me to wasting so much time since high school fruitlessly chasing the wrong ideas for over half a decade.

But ultimately character is not something that you can train. If you get hurt - if you just stand and take punches, you character will weaken. The way to strenghten it is to go on the offensive.

The way to train character is to make the right moves. You train it by winning, and by holding the right views and ideas.

9:45am. You do train it as well, by shrugging off damage. But you need to be on the right path to do that.

...

Stocks are just so great. Compare stocks to currencies and commodities on the monthly timescales and you get the feeling how powerful their upward bias is. Stock traders aren't morons for buying and holding something that persistently goes up instead of wasting their time and money on instruments that seesaw.

I recognize this. I recognize their power.

I really doubt I'll get a chance to redeem myself in this arena.

Once I get sponsors for Spiral, things will move quickly and I will be busy. If I make my riches through RL agents, by that point I will already be at the entry point to the Singularity. As great at stocks are, the benefits of being able to trade them are nowhere as great as the benefit of becoming a post human existence. It is farcial to compare them in the same breath.

9:50am. Trading thought me how to speculate. Even though my character and inexperience prevented me from making gains, they did not prevent my imagination from producing ideas. It is this effort that made me look upwards from the ground in all earnesty. My character might have been too weak to win, but if it had been too weak to even try to challenge the markets, Simulacrum would never have been written. And I would never have returned to programming.

9:55am. The best revenge against my past failures is to take the long term view and just put in effort every day towards my stated goal. I trully get a good feeling now when I think about the market, like when I finish planning out a tough part of the compiler and it all sets in.

This desire will have to be satisfied not through trading, but through programming somehow.

I'll win either way. Humans live till 80 these days. That is a long time to be able to work, and that is just if present day conditions hold. If technology development goes as planned and the Singularity happens, my expected lifespan will become indefinite. I wasted so much of my life, but what are a few decades to somebody who might live forever?

Without a doubt, I will find a way during this long life to take advantage of my experience.

10am. Let me chill. When it is time to start, I'll do so by studying those strongly connected components algorithms. I won't focus on programming today. Today studying is on the menu."