"10:05am. Did a reply on SO. I am up. Reverend Insanity is again starting to cut into my sleep. Nevermind that for now.
I made some progress thinking about the problem. I realized that you can condense the strong components to a single node in order to simplify your thinking. Instead of thinking of a circle, think of it as a single vertex.

...Ah, I see.

Suppose I had something like

A C D F
 B   E
   G

Suppose B is connected to A and C, and E is connected to D and F. And G is connected do B and E.

Then if the arrows were pointing upward, the post order would be...

A C B D F E G.

I've been thinking of the difference between the reverse post order of the reverse graph and the post order of the original. An idea came to me that even though I've been thinking of it as a sort, the reverse of the reverse post order might not necessarily be a sort. As a basic requirement, if the sort is sound, then the root of the tree should come last.

In the above example A C D F are the roots in the reverse graph, but if you look at the post order, you see that B comes in between them...

If I did the post order of the reverse graph...

G B A C E D F

Now, E comes inbetween.

...No I do not get it. Both order should be fine for the sake of separating the graph in strong components.

I said that you can think of the strong component as a single node. But that is not necessarily true. Suppose I had something like...

A C
B D E

I should group the above graph as...

(AB) (CD) E

But there is a little snag. Because of cyclical connections, it is not obvious which node come first and which last.

For example suppose I start at B and go to A. That would mean A gets added in first position. But if I start at A, then the A would definitely be added last.

Suppose I have a graph like

A B C

With arrows going left to right. If A has subnodes with cycles in it, that means that DFS starting at A instead of having the usual post order of `C B A`, could have `A C B`.

The way I am thinking of it now, a cycle means that the node might add itself ahead of turn.

I do not get why getting the reverse post order of the reverse graph would get around those difficulties.

10:40am. I am still thinking about this. Let me chill for a bit. I am going to have to code up Kosaraju and investigate it in depth. i just do not understand how it could possibly work.

10:45am. All the philosophy threads due to Elaina were wholly outside my expectations.

'Slavery is wrong, because what is somebody enslaves you?'

Well, you better work hard to attain the power so that does not happen. That is the only truth.

11:10am. Holy shit, I spent so much time writing a reply that the Elaina thread got archived.

Let me stop here so I can have dinner.

```
>>210427818
I liked the one where the common people realized the witches were exploiting them and rose up in revolt to snap their wands, turn them into rape slaves, and feed them to flowers. The good news is that no moral principles were broken in the process of doing these things.
```

```
>>210428014
I liked that part where withes murdered them all.
```

Ordinarily I would not be interested in Elaina at all - I found Kino to be a snoozefest, but the threads will make this worth it.

...

Ok, for today, my mission will be to implement Kosaraju and figure out why it works. During the night and now I generated a bunch of hypotheses, but if anything the loose reasoning is just making me more sure it would not work. I am really amazed that somebody managed to figure it out. Some algorithms like CFR are simply magical. I have no idea where inspiration for them comes from. Were those guys just trying out random things or did they have some reason for thinking it would work?

Let me get breakfast here."